homework

+++ Rays_music_exchange_homework/behaviours/Sellable.java
@@ -2,5 +2,5 @@ package behaviours;
 // Interface
 
 public interface Sellable {
-  int  cost();
+  Double  cost();
 }
